examples/full_stack_app/api/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
pyjwt>=2.8.0
werkzeug>=3.0.0
